removed everything except the framework for selecting translation APIs and the translation recipe
